requirements.txt
streamlit
google-generativeai
sentence-transformers
faiss-cpu
requests